V1.2.2753 rev713 decompiled Someone decided to alphabetically sort namespaces after 7 years of alpha dev. And 2 years after official release. Nice. Sorry for the thousand(?) changes that are basically sorted namespaces.

+++ Verse/EdgeSpan.cs
@@ -49,7 +49,7 @@ namespace Verse
 			{
 				num += 17592186044416L;
 			}
-			return (ulong)((long)num + 281474976710656L * length);
+			return num + (ulong)(281474976710656L * length);
 		}
 	}
 }
